provide thin wrapper on top of crypto primitives
The cardano-crypto is quite 'old' and has a lot of legacy inherited from past decision. One of them
was the decision of keeping every key 'XPrv' encrypted in-memory which has been said useless by our
cryptographers quite a few times.
It is necessary to encrypt the key when storing it into a more persistent location, but this isn't a
concern for this library. We can therefore simplify quite a few interfaces by removing the question
of encryption passphrases altogether.

Another annoyance of 'cardano-crypto' is that it internally stores and represent XPrv as a combination of:
  prv <> pub <> cc, and all conversion functions are made to and from this internal representation.
This is not very handy for in practice, we are often interested in either prv <> cc or pub <> cc. Thus,
our wrapper provide the necessary bits for correctly handling this conversion.