make 'HasNetworkDiscriminant' more type homogenous
Makes it easier to use from a caller perspective and also easier to document and define. I have also moved the
smart constructor for the network discriminant under the 'Style' modules so that we can make use of the type family
associated with a key type instead of having to expose the internal constructor 'AddressDiscrimination' and 'NetworkTag'.
This makes the public API slightly easier to read as it removes some steam off the reader / user.