Add a shell.nix with ghcjs support
Use `js-unknown-ghcjs-cabal` in the shell to build with ghcjs.

Uses https://github.com/input-output-hk/haskell.nix/pull/1089

The JS dependencies (those that we generate with emscripten) are moved into a separate package `cardano-addresses-jsbits`.  This package is not included in the `shellFor` `packages` list, so it will be built using nix.  That way the version of the `-jsbits` package used in the shell is the same as the one used when building with nix-build.

I had hoped to make it a sub library component rather than a package, but cabal would not pick up the sub library from the hackage database provided by haskell.nix (not sure if the problem was in haskell.nix or cabal).