Fixing Docker build. Fixes #137
I was unable to compile this project via docker by following the steps in the readme due to the `jsbits` dir not being accessible from the container.

Late on, after fixing that issue, I was also unable to run the project inside docker due to the lack of `libstdc++`. This fixed both of the issues for me.